requirements.txt
pymysql==1.1.1
SQLAlchemy==1.4.52
apache-airflow-providers-microsoft-mssql==2.0.1
apache-airflow-providers-http==4.11.1
apache-airflow-providers-common-sql==1.14.1
vertica-python==1.3.8